+++ src/main/java/webserver/http/request/RequestBody.java
@@ -1,4 +1,4 @@
-package webserver.http;
+package webserver.http.request;
 
 public class RequestBody {
