+++ src/main/java/webserver/http/request/RequestLine.java
@@ -1,4 +1,4 @@
-package webserver.http;
+package webserver.http.request;
 
 public class RequestLine {
