+++ src/main/java/webserver/http/request/RequestUri.java
@@ -1,4 +1,4 @@
-package webserver.http;
+package webserver.http.request;
 
 import com.github.jknack.handlebars.internal.lang3.StringUtils;
