+++ src/test/java/webserver/http/request/QueryParameterTest.java
@@ -1,9 +1,8 @@
-package webserver.http;
+package webserver.http.request;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.Map;
+import webserver.http.request.QueryParameter;
 
 import static org.assertj.core.api.Assertions.*;
